backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
celery[redis]==5.4.0
redis==5.2.1
faster-whisper==1.1.0
deep-translator==1.11.4
pydantic-settings==2.7.0
aiofiles==24.1.0
google-genai>=1.0.0
